update to assist with guide
As the basic guide is using git, The root compose file should be optimised to provide the best experience for that use case. Users with other setups can use many of the examples already in the repo that use volumes and such.

This change allows users to git clone into a folder of their choosing, then simply compose up. All needed folders will be made in this directory.